requirements.txt
django
djangorestframework
python-decouple
python-dotenv
gunicorn
requests
requests-oauthlib
Pillow
channels
websockets
uvicorn
celery

django-cors-headers

# TTS (可选，如需 TTS 功能)
# text-to-speech @ file:../geminar-toolchain/text-to-speech
